+++ CHANGELOG.md
@@ -15,11 +15,14 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* Now `nonce` attribute is correctly set in inline mode
+* Now the `nonce` attribute is correctly set in inline mode
 
 #### :house: Internal
 
 * Fixed typings in `core/data`
+* Moved to a new major version of yarn@4
+* Updated actions in CI
+* Migrated to node@20
 
 ## v4.0.0-beta.40 (2023-11-17)